refactor(sdk): extract messaging legacy primitives to _client/_messaging_legacy.py
PR 10/N of the SDK split — the LAST movement PR. Moves the eight
session-based A2A messaging primitives (pre-oneshot) out of
``cullis_sdk/client.py`` into a dedicated ``_MessagingLegacyMixin``
under ``cullis_sdk/_client/_messaging_legacy.py``.

Methods moved:
* ``send`` — encrypt + double-sign (inner+outer) → broker / egress.
* ``send_via_proxy`` — ADR-001 §10 resolve + mtls-only egress send.
* ``receive_via_proxy`` — pull pending session frames, verify
  mtls-only signatures.
* ``poll`` — long-poll a session's inbox, return decrypted
  ``list[InboxMessage]`` rows.
* ``ack_via_proxy`` / ``ack_message`` — ack a queued message
  (proxy / broker fork).
* ``decrypt_payload`` — E2E-decrypt an inbox row in place.
* ``_unwrap_egress_message`` (``@staticmethod``) — translate an egress
  poll-response row into the broker shape ``decrypt_payload``
  understands.

Movement only — byte-perfect. The crypto path is preserved verbatim:

* ``send`` keeps the ``sign_message(inner) → encrypt_for_agent →
  sign_message(outer)`` order and the 3-attempt connect-retry loop
  with ``time.sleep(2)`` reconnect delay.
* ``receive_via_proxy`` verifies the mtls-only signature BEFORE
  handing the row back; bad signatures raise ``ValueError`` rather
  than getting silently dropped (callers choose ack / audit / abort).
* ``decrypt_payload`` fails CLOSED on any inner-payload decrypt
  failure (logs ``RED``, raises ``ValueError`` — integrity violation).
* ``_unwrap_egress_message`` keeps the lazy ``import json as _json``
  inside the function body so the module surface is unchanged.

These methods are candidates for deprecation under the
``send_oneshot``-only direction (memory: ``oneshot_only_for_demo``),
but NO ``@deprecated`` decorator is added here — that lands in a
dedicated PR after the SDK split is complete.

MRO order:
``class CullisClient(_AiGatewayMixin, _AuthMixin, _DiscoveryMixin,
_EnrollmentMixin, _GuardianMixin, _MessagingLegacyMixin,
_MessagingOneshotMixin, _RfqMixin, _SessionsMixin, _WebSocketMixin)``
— ``L`` slots in position 6, between ``_GuardianMixin`` and
``_MessagingOneshotMixin`` (alphabetical: ``L < O``).

Side effects of the move:
* ``cullis_sdk/client.py`` drops from 896 → 453 lines (~50%).
* Removed now-unused module-level imports from ``client.py``:
  ``time``, ``uuid``, ``sign_message``, ``verify_signature``,
  ``encrypt_for_agent``, ``decrypt_from_agent``, ``log``, ``RED``.
  ``AgentInfo`` / ``InboxMessage`` / ``PubkeyFetchError`` /
  ``WebSocketConnection`` stay as backward-compat re-exports
  (external code may import them via ``cullis_sdk.client``).

Verification:
* ``pytest tests/ -n auto --dist=loadfile`` — **3500 passed**, 38
  skipped, 462s. 3 failures, all pre-existing flakes that pass in
  isolation: 2 postgres binding (``test_pg_session_and_signed_messages``,
  ``test_pg_nonce_replay_blocked``) + 1 connector state-contamination
  (``test_tampered_signature_returns_none`` — unrelated to the
  crypto path, tests local cookie HMAC tampering; passes in
  isolation, documented as same class as the
  ``test_mastio_version_defaults_to_dev`` flake).
* ``./sandbox/smoke.sh full`` — **PASS=25 FAIL=0 SKIP=1**. B4.7
  exercises the messaging path end-to-end (each agent received
  nonces from both peers under ADR-011 unified auth).

Follows the pattern of #702 / #706 / #711 / #713 / #716 / #718 /

#720 / #722 / #723.
